Install python as pre task
Ansible requires Python to be installed, but Python is not pre-installed
in some official Docker images.

With the help of the Ansible role "coglinev3.ansible_python", which uses
the module "ansible.builtin.raw", which does not require Python, Python
is now installed as a pre-task if required.

The official Docker images are now generally used, regardless of whether
they contain Python or not.